Add support for melee setups
This changes the CLI to now allow for setting two setups each of which
can used the halberd, bow, or staff. Note: there is not a lot of sanity
checking around this, so make sure the prayer is appropriate for the
style otherwise you'll get funny results.

Alongside this change is a rework of the Player struct to hold setups by
reference which simplified the logic for switching between setups.
Brought to you by lifetimes ;)