Add option to limit runs to a max success time
For example, a max_time=300 means any simulation taking more than 300
ticks (3 minutes) will be stopped and counted as a failure. This feature
is added mostly for parity with data generated from another Hunllef sim.
I do not expect it to be widely used.

Having the benchmarking is really interesting. I knew this would be
expensive as it adds a branch to the hot loop. From criterion, it makes
runs take ~6% longer for cases where it isn't used (max_time >> any
individual run). Benchmarking had previously revealed that a similar
option tick-eat has a penalty of ~8%.

I'm considering locking tick-eat and max-time behind an "advanced"
feature such that they're compiled out for other cases. In reality,
current performance is fast enough, but a 20% performance win is still
20%. And if redemption healing is added, it could certainly join them.