Add unit tests for Halberd
Fix logic bug when calculating the attack roll for Halberds. Due to a
shortcut the weapon stance bonus was being added to both the effective
attack and the effective strength level, resulting in an accuracy roll
that was slightly too high.

Tests are a wonderful thing! If I had not been adding these tests, I
would have been unlikely to turn up this bug. They remain important as
the replacement code is still brittle and prone to having a similar bug
in the future.