Fix max_time setting
We don't need to set player HP to 0 for a failure case because the later
logic already checks if Hunllef is dead before declaring success.
Instead this was removing a small number of cases where Hunllef was
killed as time ran out. We don't have handling for hit delay and this
isn't the right way to do it either.