Move certain settings into an "advanced" feature
The hope is that this will improve simulation throughput in simple cases
by removing options that add branches to the hot loop and are otherwise
unneeded in many cases. These settings are primarily used for simulating
speedrun kills.

This was not done cleanly, more as an MVP. Ideally, these settings would
also be compiled out from the CLI when the advanced feature is not
built. As of now, they can still be set, but will just do nothing.